correction style modale

+++ FrontEnd/scripts/modal.js
@@ -10,18 +10,16 @@ if (token) {
       const data = await response.json();
 
       generateModalGallery(data);
-    //suppression des éléments de l'api
 
-      const allTrashIcons = document.querySelectorAll(".fa-trash-can");
+      //suppresion de toute la galerie
 
-      let id = 0;
-      for (let i = 0; i < allTrashIcons.length; i++){
-        allTrashIcons[i].addEventListener("click", (e) => {
-          e.preventDefault();
-          //création nouveau tableau qui renvoie uniquement les id des travaux
-          const itemId = data.map(item => item.id);
-          id = itemId[i];
-          fetch(`http://localhost:5678/api/works/${id}`), {
+      const deleteAll = document.querySelector(".delete_gallery");
+
+      deleteAll.addEventListener("click", (e) => {
+        let confirm = window.confirm("Etes vous sûr de vouloir supprimer la galerie ?");
+
+        if (confirm) {
+          fetch(`http://localhost:5678/api/works`), {
             method: "DELETE",
             headers: {
               "Content-type": "application/json",
@@ -29,11 +27,40 @@ if (token) {
             }
           }
           document.querySelector(".modal_gallery").innerHTML = "";
-          generateModalGallery(data)
+          generateModalGallery(data);
+        }
+      })
+      //suppression des travaux
+
+      const allTrashIcons = document.querySelectorAll(".fa-trash-can");
+
+      let id = 0;
+      for (let i = 0; i < allTrashIcons.length; i++) {
+        allTrashIcons[i].addEventListener("click", (e) => {
+          e.preventDefault();
+          //création nouveau tableau qui renvoie uniquement les id des travaux
+          const itemId = data.map(item => item.id);
+          id = itemId[i];
+          let confirm = window.confirm("Etes vous sûr de vouloir supprimer le projet ?");
+          if (confirm) {
+            deleteElement();
+          }
+          async function deleteElement() {
+            await fetch(`http://localhost:5678/api/works/${id}`), {
+              method: "DELETE",
+              headers: {
+                "Content-type": "application/json",
+                "Authorization": "Bearer " + token
+              }
+            }
+            // regénération de la galerie
+            document.querySelector(".modal_gallery").innerHTML = "";
+            generateModalGallery(data);
+
+          }
         })
       }
 
- 
     } catch (error) {
       alert("Une erreur est survenue");
     }
@@ -46,11 +73,11 @@ if (token) {
 
       generateCategoriesOption(categories);
 
-      
-    }catch (error) {
+
+    } catch (error) {
 
     }
-  } 
+  }
 
   function generateCategoriesOption(categories) {
     const emptyCategoryOption = document.createElement("option");
@@ -58,16 +85,16 @@ if (token) {
     categorySelector.appendChild(emptyCategoryOption);
     categories.forEach(category => {
 
-        const categoryOption = document.createElement("option");
+      const categoryOption = document.createElement("option");
 
-        categoryOption.innerText = category.name;
-        categoryOption.setAttribute("value", category.name);
-        categoryOption.innerText = category.name;
+      categoryOption.innerText = category.name;
+      categoryOption.setAttribute("value", category.name);
+      categoryOption.innerText = category.name;
 
-        categorySelector.appendChild(categoryOption);
+      categorySelector.appendChild(categoryOption);
     })
   }
-        
+
   // Fonction pour générer la galerie de la modale
   function generateModalGallery(work) {
     work.forEach(item => {
@@ -80,7 +107,7 @@ if (token) {
 
       elementIllustration.src = item.imageUrl;
       elementIllustration.alt = item.title;
-      
+
 
       galleryElement.classList.add("modal_img_wrapper");
       arrowsIcon.classList.add("fa-solid");
@@ -95,7 +122,7 @@ if (token) {
       galleryElement.appendChild(trashIcon);
       galleryElement.appendChild(edit);
     });
-  
+
   }
 
   let modal = null;
@@ -113,8 +140,8 @@ if (token) {
 
     addWork.addEventListener("click", closeModal);
     addWork.addEventListener("click", openModal2);
-    
-       
+
+
 
   };
 
@@ -140,7 +167,7 @@ if (token) {
   document.querySelectorAll(".js-modal").forEach(a => {
     a.addEventListener("click", openModal);
   });
-//evenement pour ouvrir 2e modale
+  //evenement pour ouvrir 2e modale
   const addWork = document.querySelector(".add_work");
   let modal2 = null;
 
@@ -150,17 +177,17 @@ if (token) {
     modal2.removeAttribute("aria-hidden");
     modal2.setAttribute("aria-modal", "true");
     modal2.addEventListener("click", closeModal2);
-    const returnBack = modal2.querySelector(".return_back"); 
+    const returnBack = modal2.querySelector(".return_back");
     returnBack.addEventListener("click", closeModal2);
     modal2.querySelector(".js-modal-close").addEventListener("click", closeModal2);
     modal2.querySelector(".js-modal-stop").addEventListener("click", stopPropagation);
 
-    
+
   }
 
   const closeModal2 = function () {
     if (modal2 === null) return;
-    
+
     modal2.style.display = "none";
     modal2.setAttribute("aria-hidden", "true");
     modal2.removeAttribute("aria-modal");
@@ -169,12 +196,12 @@ if (token) {
     modal2.querySelector(".js-modal-stop").removeEventListener("click", stopPropagation);
     modal2 = null;
 
-    
+
   }
 
- 
-  
-  
+
+
+
 }
